services/core-api/requirements.txt
cached-property==1.5.1
factory-boy==2.12.0
flagsmith==2.0.1
Flask==3.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.0.5
GeoAlchemy2==0.6.3
ldap3==2.6
names==0.3.0
petl==1.6.8
psycopg2-binary==2.9.6
PyJWT==2.4.0
cryptography==42.0.2
pytest==8.0.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==0.10.3
redis==4.6.0
regex==2019.6.8
requests==2.31.0
responses==0.17.0
six==1.12.0
SQLAlchemy==1.4.51
sqlalchemy-filters==0.13.0
tuspy==0.2.4
utm==0.5.0
uWSGI==2.0.22
uwsgitop==0.11
Werkzeug==3.0.1
marshmallow_sqlalchemy==0.23.1
python-docx==1.1.0
sheet2dict==0.1.1
cerberus==1.3.4
celery[redis]==5.3.1
celery-redbeat==2.2.0
cachelib
gunicorn==19.10.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-dbapi==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-instrumentation-celery==0.43b0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-urllib3==0.43b0
Pillow==10.2.0
setuptools==65.5.1
urllib3==1.26.11